# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AAST Trainery API's",
        "description": "Trainery API documentation",
        "contact": {
            "email": "admin@aast-trainery.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Trainery Demo Server"
        }
    ],
    "paths": {
...
        "/W/student/profile/general": {
            "get": {
                "tags": [
                    "W-Student General Info"
                ],
                "summary": "Get Student general information",
                "description": "Get Student General Info",
                "operationId": "getStudentGeneralInfo",
                "responses": {
                    "200": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessOkVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "W-Student General Info"
                ],
                "summary": "Update Student General Info",
                "description": "Update Student General Info",
                "operationId": "App\\Http\\Controllers\\API\\V1\\StudentProfileController::generalInfo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentGeneralInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessAcceptedVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StudentGeneralInfoRequest": {
                "title": "Student Login Request",
                "description": "Student Login Request body data",
                "required": [
                    "name",
                    "phone_number",
                    "university",
                    "city",
                    "reg_no",
                    "gender",
                    "department_id",
                    "country",
                    "nationality",
                    "date_of_birth",
                    "start_year",
                    "end_year",
                    "gpa",
                    "period"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Student name",
                        "type": "string",
                        "example": "Full Name"
                    },
                    "phone_number": {
                        "title": "phone number",
                        "description": "Student phone number",
                        "type": "string",
                        "example": "+201000011111"
                    },
                    "university": {
                        "title": "university",
                        "description": "Student university",
                        "type": "string",
                        "example": "Arab Academy for science and technology"
                    },
                    "city": {
                        "title": "city",
                        "description": "Student city",
                        "type": "string",
                        "example": "Cairo"
                    },
                    "reg_no": {
                        "title": "registration number",
                        "description": "Student registration number",
                        "type": "string",
                        "example": "17200222"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "Student gender",
                        "type": "string",
                        "example": "male"
                    },
                    "department_id": {
                        "title": "department",
                        "description": "Student department",
                        "type": "string",
                        "example": "1"
                    },
                    "country": {
                        "title": "country",
                        "description": "Student country",
                        "type": "string",
                        "example": "Egypt"
                    },
                    "nationality": {
                        "title": "nationality",
                        "description": "Student nationality",
                        "type": "string",
                        "example": "Egyptian"
                    },
                    "date_of_birth": {
                        "title": "date of birth",
                        "description": "Student date of birth",
                        "type": "string",
                        "example": "1997-04-15"
                    },
                    "start_year": {
                        "title": "Student start year",
                        "description": "Student start year",
                        "type": "string",
                        "example": "2015"
                    },
                    "end_year": {
                        "title": "Student end year",
                        "description": "Student end year",
                        "type": "string",
                        "example": "2020"
                    },
                    "gpa": {
                        "title": "Student GPA",
                        "description": "Student GPA",
                        "type": "string",
                        "example": "3.5"
                    },
                    "period": {
                        "title": "Student period",
                        "description": "Student period",
                        "type": "string",
                        "example": "8"
                    },
                    "image": {
                        "title": "Student Image",
                        "description": "Student Image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "SuccessAcceptedVirtual": {
                "title": "success accepted 202",
                "description": "success accepted 202",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 201
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "example": "success created"
                    },
                    "response": {
                        "title": "response",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "description": "null or data sent in the request body",
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessOkVirtual": {
                "title": "success 200",
                "description": "success 200",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 200
                    },
                    "message": {
                        "title": "message",
                        "description": "Success message",
                        "type": "string",
                        "example": "success"
                    },
                    "response": {
                        "title": "response",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "description": "null or data sent in the request body",
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Response401Virtual": {
                "title": "response Error 401",
                "description": "response Error 401",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 401
                    },
                    "message": {
                        "title": "message",
                        "description": "unauthorized",
                        "type": "string",
                        "example": "unauthorized user"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Response422Virtual": {
                "title": "response validation Error 422",
                "description": "response validation Error 422",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 422
                    },
                    "message": {
                        "title": "message",
                        "description": "invalid data",
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field is required"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "<b>User Token:-<b><br>\n                                        <b>Header key:</b> <i>Authorization</i><br>\n                                        <b>Scheme:</b> <i>bearer</i><br>\n                                        <b>Example:</b> <i>Authorization: Bearer {JWT_token}</i>\n                                        ",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/api/1.0/auth/login",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}